backend/requirements.txt
psycopg[binary] == 3.2.10
numpy == 1.26.4
fastapi == 0.116.1 
uvicorn == 0.35.0 
python-dotenv == 1.1.0